Extend and add clarity to specification
This change adds clarity to the document along with a few additional changes.
First, it adds an "About This Specification" area and moves the MSON comment there.
Second, it adds clarity around the conflict that might arise if the deprecated
field is not set by an API change includes a deprecation.

It makes the following additions.

- It adds a `$ref` field so that external document might be referenced
- It adds an `announcement` field for public announcements in changelogs
- It allows the `x-changelog` field to be used at the root level

Lastly, it adds some clarity throughout the spec to help it make more sense
and give more insights on how to use it.